requirements.txt
# --------- pytorch --------- #
# torch==1.12.1
# torchvision==0.13.1
pytorch-lightning==1.9.5
torchmetrics==0.11.4

# --------- ---------------#
pandas==2.0.2
matplotlib==3.7.1
scikit-learn==1.2.2
scikit-image==0.21.0
tqdm==4.65.0
Pillow==9.5.0
opencv-python
albumentations

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils     # standardizing the project root setup
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
kaggle          # 
gdown
nltk
zipp